requirements.txt
# Core Framework
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
SQLAlchemy==2.0.23
Werkzeug==3.0.1

# Excel Processing
openpyxl==3.1.2
pandas==2.1.3

# PDF Generation
reportlab==4.0.7

# Utilities
python-dotenv==1.0.0
email-validator==2.1.0

# Algorithm Support
numpy==1.26.2
